Move some classes around
